refactor(auth): move and rm retrieveUserTokenInformation
