+++ src/external/authorizationClient.ts
@@ -17,10 +17,11 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { UserSession } from '@overture-stack/lyric';
+import { UserSession, UserSessionResult } from '@overture-stack/lyric';
+import { Request } from 'express';
 
 import { logger } from '@/common/logger.js';
-import { ActionIDsValues, UserDataResponse, UserDataResponseErrorType } from '@/common/types/auth.js';
+import { ActionIDsValues, Group, UserDataResponse, UserDataResponseErrorType } from '@/common/types/auth.js';
 import { userDataResponseSchema } from '@/common/validation/auth-validation.js';
 import { authConfig } from '@/config/authConfig.js';
 import { lyricProvider } from '@/core/provider.js';
@@ -127,4 +128,67 @@ export const hasAllowedAccess = async (
 	const result = await response.json();
 
 	return result[0];
+};
+
+export const retrieveUserTokenInformation = async (req: Request): Promise<UserSessionResult> => {
+	const token = extractAccessTokenFromHeader(req);
+
+	if (!token) {
+		return {
+			errorCode: 401,
+			errorMessage: 'Unauthorized: No Access token provided',
+		};
+	}
+
+	try {
+		// Determine the user information
+		const result = await fetchUserData(token);
+
+		return {
+			user: {
+				username: `${result.pcgl_id}`,
+				isAdmin: isAdmin(result.groups),
+				allowedWriteOrganizations: extractUserGroups(result.groups),
+			},
+		};
+	} catch (err) {
+		logger.error(`Error verifying token ${err}`);
+		return {
+			errorCode: 403,
+			errorMessage: 'Forbidden: Invalid token',
+		};
+	}
+};
+
+export const extractAccessTokenFromHeader = (req: Request): string | undefined => {
+	const authHeader = req.headers['authorization'];
+	if (!authHeader || !authHeader.startsWith('Bearer ')) {
+		return;
+	}
+
+	return authHeader.split(' ')[1];
+};
+
+/**
+ *
+ * @param groups List of groups users belongs to
+ * @returns boolean if user has admin group
+ */
+const isAdmin = (groups: Group[]): boolean => {
+	const { AUTH_GROUP_ADMIN } = authConfig;
+
+	return groups.some((val) => val.name === AUTH_GROUP_ADMIN);
+};
+
+/**
+ * @param groups List of groups user belongs to
+ * @returns array of groups (only the name)
+ */
+const extractUserGroups = (groups: Group[]): string[] => {
+	const parsedGroups: string[] = groups.reduce((acu: string[], currentGroup) => {
+		acu.push(currentGroup.name);
+		return acu;
+	}, []);
+
+	return parsedGroups;
 };
